+++ spng-benchmarks/benches/spng.rs
@@ -5,11 +5,9 @@ fn spng_decode(c: &mut Criterion) {
         let mut buf = Vec::new();
         b.iter(|| {
             let d = spng::Decoder::new(spng_benchmarks::TEST_PNG_002);
-            // the png crate changed its api so that that the OutputInfo is now returned
-            // from next_frame instead of read_info
-            let (_todo, mut reader) = d.read_info().unwrap();
+            let mut reader = d.read_info().unwrap();
             spng_benchmarks::reserve(&mut buf, reader.output_buffer_size());
-            let _todo_info = reader.next_frame(&mut buf).unwrap();
+            let _info = reader.next_frame(&mut buf).unwrap();
             black_box(reader);
         })
     });