+++ spng/tests/tests.rs
@@ -1,105 +1,78 @@
-use spng::{raw::ChunkAvail, Decoder};
-use std::io::Cursor;
+use spng::{raw::ChunkAvail, BitDepth, ColorType, Decoder};
+use std::io::{self, BufReader, Cursor, Read};
 
 static TEST_PNG_001: &[u8] = include_bytes!("test-001.png");
 static TEST_PNG_002: &[u8] = include_bytes!("test-002.png");
 
+fn check_decoder<R: Read>(
+    decoder: Decoder<R>,
+    width: u32,
+    height: u32,
+    bit_depth: BitDepth,
+    color_type: ColorType,
+) {
+    let mut reader = decoder.read_info().expect("read_info failed");
+    let info = reader.info();
+    let output_buffer_size = reader.output_buffer_size();
+    assert_eq!(width, info.width);
+    assert_eq!(height, info.height);
+    assert_eq!(bit_depth, info.bit_depth);
+    assert_eq!(color_type, info.color_type);
+    let mut out = vec![0; output_buffer_size];
+    let out_info = reader.next_frame(&mut out).expect("next_frame failed");
+    assert_eq!(output_buffer_size, out_info.buffer_size());
+    assert_eq!(info.width, out_info.width);
+    assert_eq!(info.height, out_info.height);
+    assert_eq!(info.bit_depth, out_info.bit_depth);
+    assert_eq!(info.color_type, out_info.color_type);
+}
+
 #[test]
-fn decode_001_cursor() -> Result<(), Box<dyn std::error::Error>> {
+fn decode_001_cursor() {
     let cursor = Cursor::new(TEST_PNG_001);
     let decoder = Decoder::new(cursor);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(300, out_info.width);
-    assert_eq!(300, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+
+    check_decoder(decoder, 300, 300, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]
-fn decode_001_cursor_buffered() -> Result<(), Box<dyn std::error::Error>> {
-    use std::io;
+fn decode_001_cursor_buffered() {
+    let cursor = Cursor::new(TEST_PNG_001);
+    let cursor_buffered = BufReader::new(cursor);
+    let decoder = Decoder::new(cursor_buffered);
 
-    let cursor = io::BufReader::new(Cursor::new(TEST_PNG_001));
-    let decoder = Decoder::new(cursor);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(300, out_info.width);
-    assert_eq!(300, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+    check_decoder(decoder, 300, 300, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]
-fn decode_001_slice() -> Result<(), Box<dyn std::error::Error>> {
+fn decode_001_slice() {
     let decoder = Decoder::new(TEST_PNG_001);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(300, out_info.width);
-    assert_eq!(300, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+
+    check_decoder(decoder, 300, 300, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]
-fn decode_002_cursor() -> Result<(), Box<dyn std::error::Error>> {
+fn decode_002_cursor() {
     let cursor = Cursor::new(TEST_PNG_002);
     let decoder = Decoder::new(cursor);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(380, out_info.width);
-    assert_eq!(287, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+
+    check_decoder(decoder, 380, 287, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]
-fn decode_002_cursor_buffered() -> Result<(), Box<dyn std::error::Error>> {
-    use std::io;
+fn decode_002_cursor_buffered() {
+    let cursor = Cursor::new(TEST_PNG_002);
+    let cursor_buffered = BufReader::new(cursor);
+    let decoder = Decoder::new(cursor_buffered);
 
-    let cursor = io::BufReader::new(Cursor::new(TEST_PNG_002));
-    let decoder = Decoder::new(cursor);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(380, out_info.width);
-    assert_eq!(287, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+    check_decoder(decoder, 380, 287, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]
-fn decode_002_slice() -> Result<(), Box<dyn std::error::Error>> {
+fn decode_002_slice() {
     let decoder = Decoder::new(TEST_PNG_002);
-    let (out_info, mut reader) = decoder.read_info()?;
-    let output_buffer_size = reader.output_buffer_size();
-    assert_eq!(380, out_info.width);
-    assert_eq!(287, out_info.height);
-    assert_eq!(8, out_info.bit_depth as u8);
-    assert_eq!(4, out_info.color_type.samples());
-    assert_eq!(out_info.buffer_size, output_buffer_size);
-    let mut out = vec![0; output_buffer_size];
-    reader.next_frame(&mut out)?;
-    Ok(())
+
+    check_decoder(decoder, 380, 287, BitDepth::Eight, ColorType::RGBA);
 }
 
 #[test]