+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated the `Decoder` `read_info` and `next_frame` API to match recent changes in the PNG crate (`v0.17.0`).
 
 ## [0.2.0-alpha.2] - 2021-08-28
 ### Changed